--- test/.eslintrc
@@ -1,5 +1,0 @@
-{
-  "env": {
-    "mocha": true,
-  }
-}
